Fix frame child positioning and update todo test
