+++ canopy/src/tutils/mod.rs
@@ -4,6 +4,7 @@ pub use ttree::*;
 use crate::{self as canopy};
 use crate::{
     backend::test::TestRender,
+    event::key,
     geom::{Direction, Expanse, Rect},
     path::Path,
     widgets::list::ListItem,
@@ -56,6 +57,51 @@ pub fn run(func: impl FnOnce(&mut Canopy, TestRender, ttree::R) -> Result<()>) -
     c.set_root_size(Expanse::new(100, 100), &mut root)?;
     ttree::reset_state();
     func(&mut c, tr, root)
+}
+
+/// A thin wrapper around [`Canopy`] that exposes a limited public API suitable
+/// for driving tests.
+pub struct Harness<'a> {
+    core: &'a mut Canopy,
+}
+
+impl<'a> Harness<'a> {
+    pub fn key<T>(&mut self, root: &mut dyn Node, k: T) -> Result<()>
+    where
+        T: Into<key::Key>,
+    {
+        self.core.key(root, k)
+    }
+
+    pub fn render(&mut self, r: &mut TestRender, root: &mut dyn Node) -> Result<()> {
+        self.core.render(r, root)
+    }
+
+    pub fn canopy(&mut self) -> &mut Canopy {
+        self.core
+    }
+}
+
+/// Run a function on a provided root node using the test render backend.
+///
+/// The root node must implement [`Loader`] so that command sets can be loaded
+/// for the test environment. The node is laid out with a default size before
+/// the supplied closure is executed.
+pub fn run_root<N>(
+    mut root: N,
+    func: impl FnOnce(&mut Harness<'_>, &mut TestRender, &mut N) -> Result<()>,
+) -> Result<()>
+where
+    N: Node + Loader,
+{
+    let (_, mut tr) = TestRender::create();
+    let mut c = Canopy::new();
+
+    <N as Loader>::load(&mut c);
+    c.set_root_size(Expanse::new(100, 100), &mut root)?;
+
+    let mut h = Harness { core: &mut c };
+    func(&mut h, &mut tr, &mut root)
 }
 
 pub struct DummyContext {}
